// 04-countryApp/src/app/countries/interfaces/cache-store.interface.ts
import {CountryInterface} from "./country.interface";
import {Region} from "./region.type";

export interface CacheStore {
  byCapital:   TermCountries;
  byCountries: TermCountries;
  byRegion:    RegionCountries;
}

export interface TermCountries {
  term: string;
  countries: CountryInterface[];
}

export interface RegionCountries {
  region:    Region;
  countries: CountryInterface[];
}